Refactored folder structure a bit
